+++ utils/rfcx-guardian/guardian-swm.js
@@ -1,0 +1,7 @@
+
+exports.swarm = {
+  validateIncomingMessage: function (req) {
+    // todo: need to know what headers need to be check
+    return true
+  }
+}